refactor: rename internal/sampling/strategy to internal/sampling/samplingstrategy
