+++ cmd/jaeger/internal/processors/adaptivesampling/processor.go
@@ -12,14 +12,14 @@ import (
 
 	"github.com/jaegertracing/jaeger/cmd/jaeger/internal/extension/remotesampling"
 	"github.com/jaegertracing/jaeger/internal/metrics/otelmetrics"
-	"github.com/jaegertracing/jaeger/internal/sampling/strategy"
+	"github.com/jaegertracing/jaeger/internal/sampling/samplingstrategy"
 	"github.com/jaegertracing/jaeger/plugin/sampling/strategyprovider/adaptive"
 	"github.com/jaegertracing/jaeger/storage_v2/v1adapter"
 )
 
 type traceProcessor struct {
 	config     *Config
-	aggregator strategy.Aggregator
+	aggregator samplingstrategy.Aggregator
 	telset     component.TelemetrySettings
 }
