+++ internal/sampling/samplingstrategy/empty_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2019 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package strategy
+package samplingstrategy
 
 import (
 	"testing"
