+++ internal/sampling/samplingstrategy/factory.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2018 The Jaeger Authors.
 // SPDX-License-Identifier: Apache-2.0
 
-package strategy
+package samplingstrategy
 
 import (
 	"go.uber.org/zap"
