+++ plugin/sampling/strategyprovider/static/factory_test.go
@@ -9,15 +9,15 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
 
-	ss "github.com/jaegertracing/jaeger/internal/sampling/strategy"
+	"github.com/jaegertracing/jaeger/internal/sampling/samplingstrategy"
 	"github.com/jaegertracing/jaeger/pkg/config"
 	"github.com/jaegertracing/jaeger/pkg/metrics"
 	"github.com/jaegertracing/jaeger/plugin"
 )
 
 var (
-	_ ss.Factory          = new(Factory)
-	_ plugin.Configurable = new(Factory)
+	_ samplingstrategy.Factory = new(Factory)
+	_ plugin.Configurable      = new(Factory)
 )
 
 func TestFactory(t *testing.T) {
